backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
langgraph==0.2.53
langchain-core==0.3.20
anthropic==0.40.0
pydantic==2.10.3
sse-starlette==2.1.3
aiofiles==24.1.0
duckduckgo-search==6.3.7
google-generativeai==0.8.3
httpx==0.27.2
# Blender Studio pipeline optional dependencies
# Install with: pip install mss Pillow pywin32
mss>=9.0.1
Pillow>=10.0.0
pywin32>=306; sys_platform == "win32"
